backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
razorpay==1.3.0
web3==6.14.0
eth-account==0.10.0
supabase==2.3.0
redis[asyncio]==5.0.1
anthropic==0.23.0
langchain==0.1.9
langchain-anthropic==0.1.4
langchain-community==0.0.24
pyjwt==2.8.0
pydantic[email]==2.5.3
httpx==0.26.0
